# src/main/webapp/openapi.yaml
openapi: 3.0.3
info:
  title: Soccorso Web API
  version: 1.0.0
  description: >
    Questa API permette di gestire un sistema di soccorso in situazioni di emergenza.
    Tutti possono inserire richieste di aiuto.
    Gli admin e gli operatori possono autenticarsi, validarle, e monitorare
    lo stato delle missioni create per rispondere a queste richieste. Gli admin possono anche annullare richieste, creare missioni e chiuderle valutando il successo
    dell'intervento. Tutto è gestito in memoria per fare il test dell'API.
    È disponibile anche un client web semplice per testare le funzionalità principali.
servers:
  - url: http://localhost:8080/SoccorsoWebServices/resources
    description: URL API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: Token di autenticazione UUID da includere nell’header Authorization
  schemas:
...
    Operator:
      type: object
      description: Rappresenta un operatore del sistema
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Giulia"
        surname:
          type: string
          example: "Rossi"
        status:
          type: string
          description: Stato dell'operatore (se disponibile o impegnato)
          example: FREE, BUSY
      required:
        - id
        - name
        - surname
        - status
    Mission:
      type: object
      description: Rappresenta una missione associata a una richiesta di soccorso.
      properties:
        id:
          type: integer
          format: int64
          description: Identificatore univoco della missione
          example: 101
        requestId:
          type: integer
          format: int64
          description: ID della richiesta di soccorso associata
          example: 5
        status:
          type: string
          description: Stato corrente della missione
          example: CLOSED, IN_PROGRESS
          default: IN_PROGRESS
        operatorIds:
          type: array
          description: Elenco degli ID degli operatori assegnati
          items:
            type: integer
            format: int64
          example: [1, 2, 3]
        start:
          type: string
          format: date-time
          description: Data e ora di inizio della missione
          example: "2025-06-30 12:25:48"
        end:
          type: string
          format: date-time
          description: Data e ora di fine della missione (può essere null se non conclusa)
          nullable: true
          example: "2025-06-30 18:25:48"
        successLevel:
          type: integer
          description: Livello di successo della missione (0=fallimento, 10=massimo)
          minimum: 0
          maximum: 10
          nullable: true
          example: 6
          default: null
      required:
        - requestId
...
paths:
...
  /missions:
    post:
      tags:
        - Missions
      summary: Crea una nuova missione
      description: |
        Crea una missione a partire da una richiesta validata con stato `ACTIVE`.
        Associa automaticamente tutti gli operatori liberi alla missione. Solo gli utenti autenticati possono fare questa operazione.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mission'
            example:
              RequestId: 5
      responses:
        '201':
          description: Missione creata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mission'
              example:
                id: 101
                requestId: 5
                status: CLOSED
                operatorIds:
                  - 1
                  - 2
                  - 3
                start: '2025-06-30 12:25:48'
                end: '2025-06-30 18:25:48'
                successLevel: 6
        '400':
          description: Richiesta non valida o non validata
        '401':
          description: Token assente o non valido
        '404':
          description: Richiesta non trovata o nessun operatore libero
  /missions/{id}/close:
    put:
      tags:
        - Missions
      summary: Chiude una missione esistente
      description: |
        Chiude una missione aperta, marcandola come `CLOSED` e assegnandole un livello di successo.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: successLevel
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 10
      responses:
        '204':
          description: Missione chiusa correttamente
        '400':
          description: Missione già chiusa
        '401':
          description: Token assente o non valido
        '404':
          description: Missione non trovata
  /missions/{id}:
    get:
      tags:
        - Missions
      summary: Recupera una missione per ID
      description: Restituisce i dettagli completi di una missione esistente.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Missione trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mission'
              example:
                id: 101
                requestId: 5
                status: IN_PROGRESS
                operatorIds:
                  - 5
                  - 101
                  - 45
                start: '2023-06-30 12:25:48'
                end: null
                successLevel: null
        '401':
          description: Token assente o non valido
        '404':
          description: Missione non trovata
  /operators:
    get:
      tags:
        - Operatori
      summary: Elenco operatori con filtro per stato
      description: |
        Restituisce gli operatori filtrati per stato (es. `FREE`, `OCCUPIED`).
        Il parametro `status` è obbligatorio.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: true
          schema:
            type: string
            example: FREE
      responses:
        '200':
          description: Elenco operatori restituito con successo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Operator'
                example:
                - id: 1
                  name: Giulia
                  surname: Rossi
                  status: FREE
                  
                - id: 2
                  name: Dario
                  surname: Moccia
                  status: BUSY
                  
        '400':
          description: Parametro `status` mancante o non valido
        '401':
          description: Token assente o non valido
        '404':
          description: Nessun operatore trovato con quello stato
  /operators/{id}:
    get:
      tags:
        - Operatori
      summary: Dettagli di un operatore
      description: | 
       Restituisce i dettagli dell'operatore specificato per `id` .
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Operatore trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operator'
              example:
                - id: 13
                  name: Massimo
                  surname: Neri
                  status: FREE
        '401':
          description: Token assente o non valido
        '404':
          description: Operatore non trovato
  /operators/{opId}/missions:
    get:
      tags:
        - Operatori
      summary: Missioni assegnate a un operatore
      description: |
        Restituisce tutte le missioni a cui l’operatore specificato ha partecipato.
        Richiede autenticazione.
      security:
        - BearerAuth: []
      parameters:
        - name: opId
          example: 4
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Elenco missioni trovate
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mission'
                example:
                - id: 101
                  requestId: 5
                  status: CLOSED
                  operatorIds:
                    - 1
                    - 4
                    - 3
                  start: '2025-06-30 12:25:48'
                  end: '2025-06-30 18:25:48'
                  successLevel: 7
                
                - id: 37
                  requestId: 5
                  status: CLOSED
                  operatorIds:
                    - 15
                    - 4
                    - 24
                  start: '2024-12-30 12:25:48'
                  end: '2025-06-30 18:25:48'
                  successLevel: 3
                
        '404':
          description: Operatore non trovato oppure nessuna missione trovata